requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
apscheduler==3.11.0
PyYAML==6.0.2
pytest==8.4.1
httpx==0.28.1
